l4t-extlinux-config: add support for OVERLAYS
As discussed in [1], overlay source on JP5 targets appears to come
only from the update capsule content in spi flash and not from
either the content appended to the extlinux dtb.  This means
if a slot switch occurs the devicetree overlay content used
will match the last capsule update as opposed to the content
appended to the dtb file referenced in extlinux.conf.  This
also means slot switches don't apply dtb overlay changes when
using slot switching instead of capsule update as a software
update strategy for non bootloader components.

With this patch, add support for writing the OVERLAYS value
in the extlinux.conf file.  Specifiying

```
UBOOT_EXTLINUX_FDT = "/boot/${DTBFILE}"
USE_OVERLAYS_IN_EXTLINUX = "1"
```
in your local.conf or machine config will
* Copy all .dtbo files in TEGRA_PLUGIN_MANAGER_OVERLAYS list
into the /boot directory of the rootfs when building the
l4t-launcher-extlinux.
* Apply a patch to JP5 edk2 firmware to support the OVERLAYS
feature, as this was not added until r36 versions of edk2-nvidia.
* Reference each .dtbo file in the extlinux.conf file, applying
it as a part of the ExtLinuxBoot process in L4TLauncher.

1:https://matrix.to/#/!YBfWVpJwNVtkmqVCPS:gitter.im/$OJAkyoJCMQDfiamxJGULUseBv7vgH2fJy9cVLMbhh4g?via=gitter.im&via=matrix.org&via=3dvisionlabs.com